// src/dropin/core/Context/Types/FieldType.ts
export enum FieldType {
  AUDIO = "audio",
  BOOLEAN = "boolean",
  COLLECTION = "collection",
  COLOR = "color",
  CUSTOM = "custom",
  DATE = "date",
  DOCUMENT = "document",
  EMAIL = "email",
  FILE = "file",
  IMAGE = "image",
  LANGUAGE = "language",
  LIST = "list",
  LOCATION = "location",
  MEASUREMENT = "measurement",
  NAME = "name",
  OBJECT = "object",
  PERIOD = "period",
  PHONE = "phone",
  PIPELINE = "pipeline",
  PRICE = "price",
  QUANTITY = "quantity",
  RECURRENCE = "recurrence",
  SELECTOR = "selector",
  TEXT = "text",
  URL = "url",
  VIDEO = "video",
}